Add a pre-up trigger to handle downloading the images tarball
This allows a "vagrant up" to Just Work the way people expect.

The logic for determining if/when to do a download (e.g. no sense
redownloading the same 0.5G tarball every time) is a little too hairy to
handle inside the Vagrantfile proper, so we're going to create a little helper
script.