Add a "get our images disk image" script
This script is invoked on "vagrant up"; it checks for the presence of an
existing disk image (images.vdi) and downloads one if it is not found.

The logic here is a bit... concise.  It's OK for right now, but we should
expand it if we start doing anything too fancy here.